generation 5 (clean up state logic)
When dealing with adding a post to both the bookmark and tag states, there was an issue with posts being taken out of the bookmark state when being tagged multiple times with different colors. I cleaned up the logic to make it only unadd from the bookmark state when the there are no tags.