Kaverna implement document groups (#137)
* generation 1 implement document groups

I made the bookmark button a menu feature where onClick brings down an option of tags. Clicking each any tag pushes the AITA post into the bookmark list, allowing the bookmark only feature to still display all the bookmarks.

* generation 2

messing with the colors of the icons to see if I could get the tags to correlate with the bookmark icon color

* generation 3

implemented colors to the bookmark, but when you have multiple bookmarks and change the color, they all change color, need to figure out how to make each individual one stay their color.

* clean up

* generation 4 (implement states)

I fixed the global color change problem by implemented states for each post, where when each post is bookmarked they are either in three colors, red, green, or blue as well as the bookmarked state. Depending on which state they are in says what color they are.

* generation 5 (clean up state logic)

When dealing with adding a post to both the bookmark and tag states, there was an issue with posts being taken out of the bookmark state when being tagged multiple times with different colors. I cleaned up the logic to make it only unadd from the bookmark state when the there are no tags.

* Cleaning up

* Build error resolution

The build error could not import the new action modules, so I am trying to see if adding these would allow them to do that

* Build resolve

* taggedPosts refactor

refactored the taggedPosts into one action, where all the tagged posts are stored with there color and depending on the postid is where the tag gets its color.

* Start to UI for addable tags (Implementing Groups)

This is the boiler plate code for editable tags

* Implement Groups Clean up UI

I cleaned up the UI for adding and searching tags, and allowed for a variety of colors to be chosen for the tags.

* Implement Document Groups (2)

I created a global list to capture all the inputed tags that the users create. Next step is to populate the tags menu to show the list.

* Implement Document Groups (Addable tags)

I had to switch from the menu tag to the select tag in order to map through the added userTags that are created and display them as tags.

* Implementing Document Groups (working baseline)

This implementation has a working baseline for users being able to add editable tags with specific colors and being able to filter the posts by those tags. There needs to be some ui clean up and extension of tags to the workspace.

* Implementing Documents Groups (Equality Calculator Correction)

Fixed object equality calculator for no duplicates when displaying tags and bookmarks